// src/frontend/web/react-admin/src/generated/models/convocation.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Mon API
 * Documentation de l'API de votre application.
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
/**
 * 
 * @export
 * @interface Convocation
 */
export interface Convocation {
    /**
     * 
     * @type {number}
     * @memberof Convocation
     */
    id?: any;
    /**
     * 
     * @type {string}
     * @memberof Convocation
     */
    libelle: any;
    /**
     * 
     * @type {string}
     * @memberof Convocation
     */
    dateConvocation: any;
    /**
     * 
     * @type {string}
     * @memberof Convocation
     */
    dateRdv: any;
    /**
     * 
     * @type {string}
     * @memberof Convocation
     */
    statut: any;
    /**
     * 
     * @type {number}
     * @memberof Convocation
     */
    personnelId: any;
    /**
     * 
     * @type {Personnel}
     * @memberof Convocation
     */
    personnel?: any;
    /**
     * 
     * @type {BaseModelpropertiescreatedAt}
     * @memberof Convocation
     */
    createdAt?: any;
    /**
     * 
     * @type {BaseModelpropertiesupdatedAt}
     * @memberof Convocation
     */
    updatedAt?: any;
    /**
     * 
     * @type {BaseModelpropertiesdeletedAt}
     * @memberof Convocation
     */
    deletedAt?: any;
}