// src/frontend/web/react-admin/src/generated/models/convocation-create-body.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Mon API
 * Documentation de l'API de votre application.
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
/**
 * 
 * @export
 * @interface ConvocationCreateBody
 */
export interface ConvocationCreateBody {
    /**
     * 
     * @type {string}
     * @memberof ConvocationCreateBody
     */
    libelle: any;
    /**
     * 
     * @type {string}
     * @memberof ConvocationCreateBody
     */
    dateConvocation: any;
    /**
     * 
     * @type {string}
     * @memberof ConvocationCreateBody
     */
    dateRdv: any;
    /**
     * 
     * @type {string}
     * @memberof ConvocationCreateBody
     */
    statut: any;
    /**
     * 
     * @type {number}
     * @memberof ConvocationCreateBody
     */
    personnelId: any;
    /**
     * 
     * @type {number}
     * @memberof ConvocationCreateBody
     */
    eleveId: any;
}